Fixed two priority calculation bugs

+++ Announcement.h
@@ -34,7 +34,7 @@ class Announcement {
 public:
     Prefix<> prefix;            // encoded with subnet mask
     uint32_t origin;            // origin ASN
-    double priority;            // priority assigned based upon path
+    uint32_t priority;            // priority assigned based upon path
     uint32_t received_from_asn; // ASN that sent the ann
     bool from_monitor = false;  // flag for seeded ann
     int64_t tstamp;             // timestamp from mrt file
@@ -47,7 +47,7 @@ public:
         prefix.netmask = anetmask;
         origin = aorigin;
         received_from_asn = from_asn;
-        priority = 0.0;
+        priority = 0;
         from_monitor = false;
         tstamp = timestamp;
     }
@@ -55,7 +55,7 @@ public:
     /** Priority constructor
      */
     Announcement(uint32_t aorigin, uint32_t aprefix, uint32_t anetmask,
-        double pr, uint32_t from_asn, int64_t timestamp, bool a_from_monitor = false) 
+        uint32_t pr, uint32_t from_asn, int64_t timestamp, bool a_from_monitor = false) 
         : Announcement(aorigin, aprefix, anetmask, from_asn, timestamp) { 
         priority = pr; 
         from_monitor = a_from_monitor;
